@@ -1,34 +1,34 @@
 module github.com/AdguardTeam/dnsproxy/internal/tools
 
-go 1.20
+go 1.22.3
 
 require (
 	github.com/fzipp/gocyclo v0.6.0
-	github.com/golangci/misspell v0.4.1
+	github.com/golangci/misspell v0.5.1
 	github.com/gordonklaus/ineffassign v0.1.0
-	github.com/kisielk/errcheck v1.6.3
+	github.com/kisielk/errcheck v1.7.0
 	github.com/kyoh86/looppointer v0.2.1
-	github.com/securego/gosec/v2 v2.18.2
+	github.com/securego/gosec/v2 v2.19.0
 	github.com/uudashr/gocognit v1.1.2
-	golang.org/x/tools v0.16.0
-	golang.org/x/vuln v1.0.1
-	honnef.co/go/tools v0.4.6
-	mvdan.cc/gofumpt v0.5.0
-	mvdan.cc/unparam v0.0.0-20230917202934-3ee2d22f45fb
+	golang.org/x/tools v0.20.0
+	golang.org/x/vuln v1.1.0
+	honnef.co/go/tools v0.4.7
+	mvdan.cc/gofumpt v0.6.0
+	mvdan.cc/unparam v0.0.0-20240427195214-063aff900ca1
 )
 
 require (
 	github.com/BurntSushi/toml v1.3.2 // indirect
-	github.com/ccojocar/zxcvbn-go v1.0.1 // indirect
+	github.com/ccojocar/zxcvbn-go v1.0.2 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
-	github.com/google/uuid v1.4.0 // indirect
+	github.com/google/uuid v1.6.0 // indirect
 	github.com/gookit/color v1.5.4 // indirect
 	github.com/kyoh86/nolint v0.0.1 // indirect
 	github.com/xo/terminfo v0.0.0-20220910002029-abceb7e1c41e // indirect
-	golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
-	golang.org/x/exp/typeparams v0.0.0-20231127185646-65229373498e // indirect
-	golang.org/x/mod v0.14.0 // indirect
-	golang.org/x/sync v0.5.0 // indirect
-	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/exp v0.0.0-20240325151524-a685a6edb6d8 // indirect
+	golang.org/x/exp/typeparams v0.0.0-20240416160154-fe59bbe5cc7f // indirect
+	golang.org/x/mod v0.17.0 // indirect
+	golang.org/x/sync v0.7.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
