@@ -1,6 +1,7 @@
 package proxy
 
 import (
+	"context"
 	"net"
 	"net/netip"
 	"strings"
@@ -29,14 +30,21 @@ var upstreamWithAddr = &fakeUpstream{
 }
 
 func TestServeCached(t *testing.T) {
-	// Prepare the proxy server.
-	dnsProxy := createTestProxy(t, nil)
-	dnsProxy.CacheEnabled = true // just one request per second is allowed
+	dnsProxy := mustNew(t, &Config{
+		UDPListenAddr:          []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TCPListenAddr:          []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		UpstreamConfig:         newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr),
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+		CacheEnabled:           true,
+	})
 
 	// Start listening.
-	err := dnsProxy.Start()
-	require.NoErrorf(t, err, "cannot start the DNS proxy: %s", err)
-	testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
+	ctx := context.Background()
+	err := dnsProxy.Start(ctx)
+	require.NoError(t, err)
+	testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
 
 	// Fill the cache.
 	reply := (&dns.Msg{
@@ -263,12 +271,20 @@ func TestCache_concurrent(t *testing.T) {
 }
 
 func TestCacheExpiration(t *testing.T) {
-	dnsProxy := createTestProxy(t, nil)
-	dnsProxy.CacheEnabled = true
+	dnsProxy := mustNew(t, &Config{
+		UDPListenAddr:          []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TCPListenAddr:          []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		UpstreamConfig:         newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr),
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+		CacheEnabled:           true,
+	})
 
-	err := dnsProxy.Start()
+	ctx := context.Background()
+	err := dnsProxy.Start(ctx)
 	require.NoError(t, err)
-	testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
+	testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
 
 	// Create dns messages with TTL of 1 second.
 	rrs := []dns.RR{
@@ -310,22 +326,31 @@ func TestCacheExpiration(t *testing.T) {
 }
 
 func TestCacheExpirationWithTTLOverride(t *testing.T) {
-	dnsProxy := createTestProxy(t, nil)
-	dnsProxy.CacheEnabled = true
-	dnsProxy.CacheMinTTL = 20
-	dnsProxy.CacheMaxTTL = 40
-
 	u := testUpstream{}
-	dnsProxy.UpstreamConfig.Upstreams = []upstream.Upstream{&u}
 
-	err := dnsProxy.Start()
+	dnsProxy := mustNew(t, &Config{
+		UDPListenAddr: []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TCPListenAddr: []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		UpstreamConfig: &UpstreamConfig{
+			Upstreams: []upstream.Upstream{&u},
+		},
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+		CacheEnabled:           true,
+		CacheMinTTL:            20,
+		CacheMaxTTL:            40,
+	})
+
+	ctx := context.Background()
+	err := dnsProxy.Start(ctx)
 	require.NoError(t, err)
-	testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
+	testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
 
 	d := &DNSContext{}
 
 	t.Run("replace_min", func(t *testing.T) {
-		d.Req = createHostTestMessage("host")
+		d.Req = newHostTestMessage("host")
 		d.Addr = netip.AddrPort{}
 
 		u.ans = []dns.RR{&dns.A{
@@ -349,7 +374,7 @@ func TestCacheExpirationWithTTLOverride(t *testing.T) {
 	})
 
 	t.Run("replace_max", func(t *testing.T) {
-		d.Req = createHostTestMessage("host2")
+		d.Req = newHostTestMessage("host2")
 		d.Addr = netip.AddrPort{}
 
 		u.ans = []dns.RR{&dns.A{
@@ -568,7 +593,7 @@ func setAndGetCache(t *testing.T, c *cache, g *sync.WaitGroup, host, ip string) 
 
 	c.set(dnsMsg, upstreamWithAddr)
 
-	for i := 0; i < 2; i++ {
+	for range 2 {
 		ci, expired, key := c.get(dnsMsg)
 		require.NotNilf(t, ci, "no cache found for %s", host)
 
