@@ -6,28 +6,31 @@ import (
 	"time"
 )
 
-// TODO(e.burkov):  Rewrite the cache using zero-values instead of storing
-// useless boolean as an integer.
-
 const (
-	fastestAddrCacheTTLSec = 10 * 60 // cache TTL for IP addresses
+	// fastestAddrCacheTTLSec is the cache TTL for IP addresses.
+	fastestAddrCacheTTLSec = 10 * 60
 )
 
+// cacheEntry represents an item that will be stored in the cache.
+//
+// TODO(e.burkov): Rewrite the cache using zero-values instead of storing
+// useless boolean as an integer.
 type cacheEntry struct {
-	status      int // 0:ok; 1:timed out
+	// status is 1 if the item is timed out.
+	status      int
 	latencyMsec uint
 }
 
 // packCacheEntry packs the cache entry and the TTL to bytes in the following
 // order:
 //
-//	expire   [4]byte  (Unix time, seconds)
-//	status   byte     (0 for ok, 1 for timed out)
-//	latency  [2]byte  (milliseconds)
-func packCacheEntry(ent *cacheEntry, ttl uint32) []byte {
+//   - expire   [4]byte  (Unix time, seconds),
+//   - status   byte     (0 for ok, 1 for timed out),
+//   - latency  [2]byte  (milliseconds).
+func packCacheEntry(ent *cacheEntry, ttl uint32) (d []byte) {
 	expire := uint32(time.Now().Unix()) + ttl
 
-	d := make([]byte, 4+1+2)
+	d = make([]byte, 4+1+2)
 	binary.BigEndian.PutUint32(d, expire)
 	i := 4
 
@@ -40,15 +43,16 @@ func packCacheEntry(ent *cacheEntry, ttl uint32) []byte {
 	return d
 }
 
-// unpackCacheEntry - unpacks bytes to cache entry and checks TTL
-// if the record is expired, returns nil
-func unpackCacheEntry(data []byte) *cacheEntry {
+// unpackCacheEntry unpacks bytes to cache entry and checks TTL, if the record
+// is expired returns nil.
+func unpackCacheEntry(data []byte) (ent *cacheEntry) {
 	now := time.Now().Unix()
 	expire := binary.BigEndian.Uint32(data[:4])
 	if int64(expire) <= now {
 		return nil
 	}
-	ent := cacheEntry{}
+
+	ent = &cacheEntry{}
 	i := 4
 
 	ent.status = int(data[i])
@@ -57,24 +61,21 @@ func unpackCacheEntry(data []byte) *cacheEntry {
 	ent.latencyMsec = uint(binary.BigEndian.Uint16(data[i:]))
 	// i += 2
 
-	return &ent
+	return ent
 }
 
-// cacheFind - find entry in the cache for this IP
-// returns null if nothing found or if the record for this ip is expired
-func (f *FastestAddr) cacheFind(ip netip.Addr) *cacheEntry {
+// cacheFind finds entry in the cache for the given IP address.  Returns nil if
+// nothing is found or if the record is expired.
+func (f *FastestAddr) cacheFind(ip netip.Addr) (ent *cacheEntry) {
 	val := f.ipCache.Get(ip.AsSlice())
 	if val == nil {
 		return nil
 	}
-	ent := unpackCacheEntry(val)
-	if ent == nil {
-		return nil
-	}
-	return ent
+
+	return unpackCacheEntry(val)
 }
 
-// cacheAddFailure - store unsuccessful attempt in cache
+// cacheAddFailure stores unsuccessful attempt in cache.
 func (f *FastestAddr) cacheAddFailure(ip netip.Addr) {
 	ent := cacheEntry{
 		status: 1,
@@ -88,8 +89,8 @@ func (f *FastestAddr) cacheAddFailure(ip netip.Addr) {
 	}
 }
 
-// store a successful ping result in cache
-// replace previous result if our latency is lower
+// cacheAddSuccessful stores a successful ping result in the cache.  Replaces
+// previous result if our latency is lower.
 func (f *FastestAddr) cacheAddSuccessful(ip netip.Addr, latency uint) {
 	ent := cacheEntry{
 		latencyMsec: latency,
@@ -104,7 +105,7 @@ func (f *FastestAddr) cacheAddSuccessful(ip netip.Addr, latency uint) {
 	}
 }
 
-// cacheAdd -- adds a new entry to the cache
+// cacheAdd adds a new entry to the cache.
 func (f *FastestAddr) cacheAdd(ent *cacheEntry, ip netip.Addr, ttl uint32) {
 	val := packCacheEntry(ent, ttl)
 	f.ipCache.Set(ip.AsSlice(), val)
