@@ -12,10 +12,10 @@ import (
 
 	"github.com/AdguardTeam/dnsproxy/upstream"
 	"github.com/AdguardTeam/golibs/cache"
+	"github.com/AdguardTeam/golibs/container"
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/miekg/dns"
-	"golang.org/x/exp/maps"
 )
 
 // DefaultPingWaitTimeout is the default period of time for waiting ping
@@ -24,8 +24,7 @@ const DefaultPingWaitTimeout = 1 * time.Second
 
 // FastestAddr provides methods to determine the fastest network addresses.
 type FastestAddr struct {
-	// pinger is the dialer with predefined timeout for pinging TCP
-	// connections.
+	// pinger is the dialer with predefined timeout for pinging TCP connections.
 	pinger *net.Dialer
 
 	// ipCacheLock protects ipCache.
@@ -37,15 +36,14 @@ type FastestAddr struct {
 	// pingPorts are the ports to ping on.
 	pingPorts []uint
 
-	// PingWaitTimeout is the timeout for waiting all the resolved addresses
-	// are pinged.  Any ping results received after it are cached but not
-	// used at the moment.  It should be configured right after the
-	// FastestAddr initialization since it isn't protected for concurrent
-	// usage.
+	// PingWaitTimeout is the timeout for waiting all the resolved addresses to
+	// be pinged.  Any ping results received after that moment are cached, but
+	// won't be used.  It should be configured right after the FastestAddr
+	// initialization since it isn't protected for concurrent usage.
 	PingWaitTimeout time.Duration
 }
 
-// NewFastestAddr initializes a new instance of the *FastestAddr.
+// NewFastestAddr initializes a new instance of *FastestAddr.
 func NewFastestAddr() (f *FastestAddr) {
 	return &FastestAddr{
 		ipCacheLock: &sync.Mutex{},
@@ -59,43 +57,41 @@ func NewFastestAddr() (f *FastestAddr) {
 	}
 }
 
-// ExchangeFastest queries each specified upstream and returns a response with
+// ExchangeFastest queries each specified upstream and returns the response with
 // the fastest IP address.  The fastest IP address is considered to be the first
 // one successfully dialed and other addresses are removed from the answer.
-func (f *FastestAddr) ExchangeFastest(req *dns.Msg, ups []upstream.Upstream) (
-	resp *dns.Msg,
-	u upstream.Upstream,
-	err error,
-) {
+func (f *FastestAddr) ExchangeFastest(
+	req *dns.Msg,
+	ups []upstream.Upstream,
+) (resp *dns.Msg, u upstream.Upstream, err error) {
 	replies, err := upstream.ExchangeAll(ups, req)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	host := strings.ToLower(req.Question[0].Name)
-
-	ipSet := map[netip.Addr]struct{}{}
+	ipSet := container.NewMapSet[netip.Addr]()
 	for _, r := range replies {
 		for _, rr := range r.Resp.Answer {
 			ip := ipFromRR(rr)
-			if _, ok := ipSet[ip]; !ok && ip != (netip.Addr{}) {
-				ipSet[ip] = struct{}{}
+			if ip.IsValid() && !ip.IsUnspecified() {
+				ipSet.Add(ip)
 			}
 		}
 	}
 
-	ips := maps.Keys(ipSet)
+	ips := ipSet.Values()
+	host := strings.ToLower(req.Question[0].Name)
 	if pingRes := f.pingAll(host, ips); pingRes != nil {
 		return f.prepareReply(pingRes, replies)
 	}
 
-	log.Debug("%s: no fastest IP found, using the first response", host)
+	log.Debug("fastip: %s: no fastest IP found, using the first response", host)
 
 	return replies[0].Resp, replies[0].Upstream, nil
 }
 
 // prepareReply converts replies into the DNS answer message according to res.
-// The returned upstreams is the one which replied with the fastest address.
+// The returned upstream is the one which replied with the fastest address.
 func (f *FastestAddr) prepareReply(
 	res *pingResult,
 	replies []upstream.ExchangeAllResult,
@@ -111,13 +107,20 @@ func (f *FastestAddr) prepareReply(
 	}
 
 	if resp == nil {
-		log.Error("found no replies with IP %s, most likely this is a bug", ip)
+		log.Error("fastip: found no replies with IP %s, most likely this is a bug", ip)
 
+		// TODO(d.kolyshev): Consider returning error?
 		return replies[0].Resp, replies[0].Upstream, nil
 	}
 
-	// Modify the message and keep only A and AAAA records containing the
-	// fastest IP address.
+	filterResponseAnswer(resp, ip)
+
+	return resp, u, nil
+}
+
+// filterResponseAnswer modifies the response message, it keeps only A and AAAA
+// records with the given IP address.
+func filterResponseAnswer(resp *dns.Msg, ip netip.Addr) {
 	ans := make([]dns.RR, 0, len(resp.Answer))
 	ipBytes := ip.AsSlice()
 	for _, rr := range resp.Answer {
@@ -137,8 +140,6 @@ func (f *FastestAddr) prepareReply(
 
 	// Set new answer.
 	resp.Answer = ans
-
-	return resp, u, nil
 }
 
 // hasInAns returns true if m contains ip in its Answer section.
