@@ -2,6 +2,7 @@
 package main
 
 import (
+	"context"
 	"crypto/tls"
 	"fmt"
 	"github.com/ameshkov/dnscrypt/v2"
@@ -25,12 +26,12 @@ import (
 	"github.com/AdguardTeam/dnsproxy/proxy"
 	"github.com/AdguardTeam/dnsproxy/upstream"
 	"github.com/AdguardTeam/golibs/log"
-	"github.com/AdguardTeam/golibs/mathutil"
 	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/AdguardTeam/golibs/osutil"
 	"github.com/AdguardTeam/golibs/timeutil"
-	"github.com/go-co-op/gocron"
+	"github.com/ameshkov/dnscrypt/v2"
 	goFlags "github.com/jessevdk/go-flags"
+	"gopkg.in/yaml.v3"
 )
 
 // Options represents console arguments.  For further additions, please do not
@@ -44,147 +45,16 @@ type Options struct {
 	// options.
 	ConfigPath string `long:"config-path" description:"yaml configuration file. Minimal working configuration in config.yaml.dist. Options passed through command line will override the ones from this file." default:""`
 
-	// Log settings
-	// --
-
-	// Should we write
-	Verbose bool `yaml:"verbose" short:"v" long:"verbose" description:"Verbose output (optional)" optional:"yes" optional-value:"true"`
-
-	// Path to a log file
+	// LogOutput is the path to the log file.
 	LogOutput string `yaml:"output" short:"o" long:"output" description:"Path to the log file. If not set, write to stdout."`
 
-	// Listen addrs
-	// --
-
-	// Server listen address
-	ListenAddrs []string `yaml:"listen-addrs" short:"l" long:"listen" description:"Listening addresses"`
-
-	// Server listen ports
-	ListenPorts []int `yaml:"listen-ports" short:"p" long:"port" description:"Listening ports. Zero value disables TCP and UDP listeners"`
-
-	// HTTPS listen ports
-	HTTPSListenPorts []int `yaml:"https-port" short:"s" long:"https-port" description:"Listening ports for DNS-over-HTTPS"`
-
-	// TLS listen ports
-	TLSListenPorts []int `yaml:"tls-port" short:"t" long:"tls-port" description:"Listening ports for DNS-over-TLS"`
-
-	// QUIC listen ports
-	QUICListenPorts []int `yaml:"quic-port" short:"q" long:"quic-port" description:"Listening ports for DNS-over-QUIC"`
-
-	// DNSCrypt listen ports
-	DNSCryptListenPorts []int `yaml:"dnscrypt-port" short:"y" long:"dnscrypt-port" description:"Listening ports for DNSCrypt"`
-
-	// Encryption config
-	// --
-
-	// Path to the .crt with the certificate chain
+	// TLSCertPath is the path to the .crt with the certificate chain.
 	TLSCertPath string `yaml:"tls-crt" short:"c" long:"tls-crt" description:"Path to a file with the certificate chain"`
 
-	// Path to the file with the private key
+	// TLSKeyPath is the path to the file with the private key.
 	TLSKeyPath string `yaml:"tls-key" short:"k" long:"tls-key" description:"Path to a file with the private key"`
 
-	// Minimum TLS version
-	TLSMinVersion float32 `yaml:"tls-min-version" long:"tls-min-version" description:"Minimum TLS version, for example 1.0" optional:"yes"`
-
-	// Maximum TLS version
-	TLSMaxVersion float32 `yaml:"tls-max-version" long:"tls-max-version" description:"Maximum TLS version, for example 1.3" optional:"yes"`
-
-	// Disable TLS certificate verification
-	Insecure bool `yaml:"insecure" long:"insecure" description:"Disable secure TLS certificate validation" optional:"yes" optional-value:"false"`
-
-	// Path to the DNSCrypt configuration file
-	DNSCryptConfigPath string `yaml:"dnscrypt-config" short:"g" long:"dnscrypt-config" description:"Path to a file with DNSCrypt configuration. You can generate one using https://github.com/ameshkov/dnscrypt"`
-
-	// HTTP3 controls whether HTTP/3 is enabled for this instance of dnsproxy.
-	// It enables HTTP/3 support for both the DoH upstreams and the DoH server.
-	HTTP3 bool `yaml:"http3" long:"http3" description:"Enable HTTP/3 support" optional:"yes" optional-value:"false"`
-
-	// Upstream DNS servers settings
-	// --
-
-	// DNS upstreams
-	Upstreams []string `yaml:"upstream" short:"u" long:"upstream" description:"An upstream to be used (can be specified multiple times). You can also specify path to a file with the list of servers" optional:"false"`
-
-	// Bootstrap DNS
-	BootstrapDNS []string `yaml:"bootstrap" short:"b" long:"bootstrap" description:"Bootstrap DNS for DoH and DoT, can be specified multiple times (default: use system-provided)"`
-
-	// Fallback DNS resolver
-	Fallbacks []string `yaml:"fallback" short:"f" long:"fallback" description:"Fallback resolvers to use when regular ones are unavailable, can be specified multiple times. You can also specify path to a file with the list of servers"`
-
-	// PrivateRDNSUpstreams are upstreams to use for reverse DNS lookups of
-	// private addresses.
-	PrivateRDNSUpstreams []string `yaml:"private-rdns-upstream" long:"private-rdns-upstream" description:"Private DNS upstreams to use for reverse DNS lookups of private addresses, can be specified multiple times"`
-
-	// If true, parallel queries to all configured upstream servers
-	AllServers bool `yaml:"all-servers" long:"all-servers" description:"If specified, parallel queries to all configured upstream servers are enabled" optional:"yes" optional-value:"true"`
-
-	// Respond to A or AAAA requests only with the fastest IP address
-	//  detected by ICMP response time or TCP connection time
-	FastestAddress bool `yaml:"fastest-addr" long:"fastest-addr" description:"Respond to A or AAAA requests only with the fastest IP address" optional:"yes" optional-value:"true"`
-
-	// Timeout for outbound DNS queries to remote upstream servers in a
-	// human-readable form.  Default is 10s.
-	Timeout timeutil.Duration `yaml:"timeout" long:"timeout" description:"Timeout for outbound DNS queries to remote upstream servers in a human-readable form" default:"10s"`
-
-	// Cache settings
-	// --
-
-	// If true, DNS cache is enabled
-	Cache bool `yaml:"cache" long:"cache" description:"If specified, DNS cache is enabled" optional:"yes" optional-value:"true"`
-
-	// Cache size value
-	CacheSizeBytes int `yaml:"cache-size" long:"cache-size" description:"Cache size (in bytes). Default: 64k"`
-
-	// DNS cache minimum TTL value - overrides record value
-	CacheMinTTL uint32 `yaml:"cache-min-ttl" long:"cache-min-ttl" description:"Minimum TTL value for DNS entries, in seconds. Capped at 3600. Artificially extending TTLs should only be done with careful consideration."`
-
-	// DNS cache maximum TTL value - overrides record value
-	CacheMaxTTL uint32 `yaml:"cache-max-ttl" long:"cache-max-ttl" description:"Maximum TTL value for DNS entries, in seconds."`
-
-	// CacheOptimistic, if set to true, enables the optimistic DNS cache. That means that cached results will be served even if their cache TTL has already expired.
-	CacheOptimistic bool `yaml:"cache-optimistic" long:"cache-optimistic" description:"If specified, optimistic DNS cache is enabled" optional:"yes" optional-value:"true"`
-
-	// Anti-DNS amplification measures
-	// --
-
-	// Ratelimit value
-	Ratelimit int `yaml:"ratelimit" short:"r" long:"ratelimit" description:"Ratelimit (requests per second)"`
-
-	// RatelimitSubnetLenIPv4 is a subnet length for IPv4 addresses used for
-	// rate limiting requests
-	RatelimitSubnetLenIPv4 int `yaml:"ratelimit-subnet-len-ipv4" long:"ratelimit-subnet-len-ipv4" description:"Ratelimit subnet length for IPv4." default:"24"`
-
-	// RatelimitSubnetLenIPv6 is a subnet length for IPv6 addresses used for
-	// rate limiting requests
-	RatelimitSubnetLenIPv6 int `yaml:"ratelimit-subnet-len-ipv6" long:"ratelimit-subnet-len-ipv6" description:"Ratelimit subnet length for IPv6." default:"56"`
-
-	// If true, refuse ANY requests
-	RefuseAny bool `yaml:"refuse-any" long:"refuse-any" description:"If specified, refuse ANY requests" optional:"yes" optional-value:"true"`
-
-	// ECS settings
-	// --
-
-	// Use EDNS Client Subnet extension
-	EnableEDNSSubnet bool `yaml:"edns" long:"edns" description:"Use EDNS Client Subnet extension" optional:"yes" optional-value:"true"`
-
-	// Use Custom EDNS Client Address
-	EDNSAddr string `yaml:"edns-addr" long:"edns-addr" description:"Send EDNS Client Address"`
-
-	// DNS64 settings
-	// --
-
-	// Defines whether DNS64 functionality is enabled or not
-	DNS64 bool `yaml:"dns64" long:"dns64" description:"If specified, dnsproxy will act as a DNS64 server" optional:"yes" optional-value:"true"`
-
-	// DNS64Prefix defines the DNS64 prefixes that dnsproxy should use when it
-	// acts as a DNS64 server.  If not specified, dnsproxy uses the default
-	// Well-Known Prefix.  This option can be specified multiple times.
-	DNS64Prefix []string `yaml:"dns64-prefix" long:"dns64-prefix" description:"Prefix used to handle DNS64. If not specified, dnsproxy uses the 'Well-Known Prefix' 64:ff9b::.  Can be specified multiple times" required:"false"`
-
-	// Other settings and options
-	// --
-
-	// Set Server header for the HTTPS server
+	// HTTPSServerName sets Server header for the HTTPS server.
 	HTTPSServerName string `yaml:"https-server-name" long:"https-server-name" description:"Set the Server header for the responses from the HTTPS server." default:"dnsproxy"`
 
 	// HTTPSUserinfo is the sole permitted userinfo for the DoH basic
@@ -192,23 +62,105 @@ type Options struct {
 	// basic authentication information.
 	HTTPSUserinfo string `yaml:"https-userinfo" long:"https-userinfo" description:"If set, all DoH queries are required to have this basic authentication information."`
 
-	// If true, all AAAA requests will be replied with NoError RCode and empty answer
-	IPv6Disabled bool `yaml:"ipv6-disabled" long:"ipv6-disabled" description:"If specified, all AAAA requests will be replied with NoError RCode and empty answer" optional:"yes" optional-value:"true"`
+	// DNSCryptConfigPath is the path to the DNSCrypt configuration file.
+	DNSCryptConfigPath string `yaml:"dnscrypt-config" short:"g" long:"dnscrypt-config" description:"Path to a file with DNSCrypt configuration. You can generate one using https://github.com/ameshkov/dnscrypt"`
 
-	// Transform responses that contain at least one of the given IP addresses into NXDOMAIN
+	// EDNSAddr is the custom EDNS Client Address to send.
+	EDNSAddr string `yaml:"edns-addr" long:"edns-addr" description:"Send EDNS Client Address"`
+
+	// ListenAddrs is the list of server's listen addresses.
+	ListenAddrs []string `yaml:"listen-addrs" short:"l" long:"listen" description:"Listening addresses"`
+
+	// ListenPorts are the ports server listens on.
+	ListenPorts []int `yaml:"listen-ports" short:"p" long:"port" description:"Listening ports. Zero value disables TCP and UDP listeners"`
+
+	// HTTPSListenPorts are the ports server listens on for DNS-over-HTTPS.
+	HTTPSListenPorts []int `yaml:"https-port" short:"s" long:"https-port" description:"Listening ports for DNS-over-HTTPS"`
+
+	// TLSListenPorts are the ports server listens on for DNS-over-TLS.
+	TLSListenPorts []int `yaml:"tls-port" short:"t" long:"tls-port" description:"Listening ports for DNS-over-TLS"`
+
+	// QUICListenPorts are the ports server listens on for DNS-over-QUIC.
+	QUICListenPorts []int `yaml:"quic-port" short:"q" long:"quic-port" description:"Listening ports for DNS-over-QUIC"`
+
+	// DNSCryptListenPorts are the ports server listens on for DNSCrypt.
+	DNSCryptListenPorts []int `yaml:"dnscrypt-port" short:"y" long:"dnscrypt-port" description:"Listening ports for DNSCrypt"`
+
+	// Upstreams is the list of DNS upstream servers.
+	Upstreams []string `yaml:"upstream" short:"u" long:"upstream" description:"An upstream to be used (can be specified multiple times). You can also specify path to a file with the list of servers" optional:"false"`
+
+	// BootstrapDNS is the list of bootstrap DNS upstream servers.
+	BootstrapDNS []string `yaml:"bootstrap" short:"b" long:"bootstrap" description:"Bootstrap DNS for DoH and DoT, can be specified multiple times (default: use system-provided)"`
+
+	// Fallbacks is the list of fallback DNS upstream servers.
+	Fallbacks []string `yaml:"fallback" short:"f" long:"fallback" description:"Fallback resolvers to use when regular ones are unavailable, can be specified multiple times. You can also specify path to a file with the list of servers"`
+
+	// PrivateRDNSUpstreams are upstreams to use for reverse DNS lookups of
+	// private addresses, including the requests for authority records, such as
+	// SOA and NS.
+	PrivateRDNSUpstreams []string `yaml:"private-rdns-upstream" long:"private-rdns-upstream" description:"Private DNS upstreams to use for reverse DNS lookups of private addresses, can be specified multiple times"`
+
+	// DNS64Prefix defines the DNS64 prefixes that dnsproxy should use when it
+	// acts as a DNS64 server.  If not specified, dnsproxy uses the default
+	// Well-Known Prefix.  This option can be specified multiple times.
+	DNS64Prefix []string `yaml:"dns64-prefix" long:"dns64-prefix" description:"Prefix used to handle DNS64. If not specified, dnsproxy uses the 'Well-Known Prefix' 64:ff9b::.  Can be specified multiple times" required:"false"`
+
+	// PrivateSubnets is the list of private subnets to determine private
+	// addresses.
+	PrivateSubnets []string `yaml:"private-subnets" long:"private-subnets" description:"Private subnets to use for reverse DNS lookups of private addresses" required:"false"`
+
+	// BogusNXDomain transforms responses that contain at least one of the given
+	// IP addresses into NXDOMAIN.
+	//
+	// TODO(a.garipov): Find a way to use [netutil.Prefix].  Currently, package
+	// go-flags doesn't support text unmarshalers.
 	BogusNXDomain []string `yaml:"bogus-nxdomain" long:"bogus-nxdomain" description:"Transform the responses containing at least a single IP that matches specified addresses and CIDRs into NXDOMAIN.  Can be specified multiple times."`
 
-	// UDP buffer size value
+	// Timeout for outbound DNS queries to remote upstream servers in a
+	// human-readable form.  Default is 10s.
+	Timeout timeutil.Duration `yaml:"timeout" long:"timeout" description:"Timeout for outbound DNS queries to remote upstream servers in a human-readable form" default:"10s"`
+
+	// CacheMinTTL is the minimum TTL value for caching DNS entries, in seconds.
+	// It overrides the TTL value from the upstream server, if the one is less.
+	CacheMinTTL uint32 `yaml:"cache-min-ttl" long:"cache-min-ttl" description:"Minimum TTL value for DNS entries, in seconds. Capped at 3600. Artificially extending TTLs should only be done with careful consideration."`
+
+	// CacheMaxTTL is the maximum TTL value for caching DNS entries, in seconds.
+	// It overrides the TTL value from the upstream server, if the one is
+	// greater.
+	CacheMaxTTL uint32 `yaml:"cache-max-ttl" long:"cache-max-ttl" description:"Maximum TTL value for DNS entries, in seconds."`
+
+	// CacheSizeBytes is the cache size in bytes.  Default is 64k.
+	CacheSizeBytes int `yaml:"cache-size" long:"cache-size" description:"Cache size (in bytes). Default: 64k"`
+
+	// Ratelimit is the maximum number of requests per second.
+	Ratelimit int `yaml:"ratelimit" short:"r" long:"ratelimit" description:"Ratelimit (requests per second)"`
+
+	// RatelimitSubnetLenIPv4 is a subnet length for IPv4 addresses used for
+	// rate limiting requests.
+	RatelimitSubnetLenIPv4 int `yaml:"ratelimit-subnet-len-ipv4" long:"ratelimit-subnet-len-ipv4" description:"Ratelimit subnet length for IPv4." default:"24"`
+
+	// RatelimitSubnetLenIPv6 is a subnet length for IPv6 addresses used for
+	// rate limiting requests.
+	RatelimitSubnetLenIPv6 int `yaml:"ratelimit-subnet-len-ipv6" long:"ratelimit-subnet-len-ipv6" description:"Ratelimit subnet length for IPv6." default:"56"`
+
+	// UDPBufferSize is the size of the UDP buffer in bytes.  A value <= 0 will
+	// use the system default.
 	UDPBufferSize int `yaml:"udp-buf-size" long:"udp-buf-size" description:"Set the size of the UDP buffer in bytes. A value <= 0 will use the system default."`
 
 	// MaxGoRoutines is the maximum number of goroutines.
 	MaxGoRoutines uint `yaml:"max-go-routines" long:"max-go-routines" description:"Set the maximum number of go routines. A zero value will not not set a maximum."`
 
+	// TLSMinVersion is the minimum allowed version of TLS.
+	TLSMinVersion float32 `yaml:"tls-min-version" long:"tls-min-version" description:"Minimum TLS version, for example 1.0" optional:"yes"`
+
+	// TLSMaxVersion is the maximum allowed version of TLS.
+	TLSMaxVersion float32 `yaml:"tls-max-version" long:"tls-max-version" description:"Maximum TLS version, for example 1.3" optional:"yes"`
+
 	// Pprof defines whether the pprof information needs to be exposed via
 	// localhost:6060 or not.
 	Pprof bool `yaml:"pprof" long:"pprof" description:"If present, exposes pprof information on localhost:6060." optional:"yes" optional-value:"true"`
 
-	// Print DNSProxy version (just for the help)
+	// Version, if true, prints the program version, and exits.
 	Version bool `yaml:"version" long:"version" description:"Prints the program version"`
 
 	// rafal code
@@ -222,6 +174,50 @@ type Options struct {
 	ExcludedFromCachingLists []string `yaml:"domains_excluded_from_caching" long:"domains_excluded_from_caching" description:"The list of domains to be excluded from caching (can be specified multiple times)."`
 	///////////////////////////////////////////////////////////////////////////////
 	// end rafal code
+
+	// Verbose controls the verbosity of the output.
+	Verbose bool `yaml:"verbose" short:"v" long:"verbose" description:"Verbose output (optional)" optional:"yes" optional-value:"true"`
+
+	// Insecure disables upstream servers TLS certificate verification.
+	Insecure bool `yaml:"insecure" long:"insecure" description:"Disable secure TLS certificate validation" optional:"yes" optional-value:"false"`
+
+	// IPv6Disabled makes the server to respond with NODATA to all AAAA queries.
+	IPv6Disabled bool `yaml:"ipv6-disabled" long:"ipv6-disabled" description:"If specified, all AAAA requests will be replied with NoError RCode and empty answer" optional:"yes" optional-value:"true"`
+
+	// HTTP3 controls whether HTTP/3 is enabled for this instance of dnsproxy.
+	// It enables HTTP/3 support for both the DoH upstreams and the DoH server.
+	HTTP3 bool `yaml:"http3" long:"http3" description:"Enable HTTP/3 support" optional:"yes" optional-value:"false"`
+
+	// AllServers makes server to query all configured upstream servers in
+	// parallel.
+	AllServers bool `yaml:"all-servers" long:"all-servers" description:"If specified, parallel queries to all configured upstream servers are enabled" optional:"yes" optional-value:"true"`
+
+	// FastestAddress controls whether the server should respond to A or AAAA
+	// requests only with the fastest IP address detected by ICMP response time
+	// or TCP connection time.
+	FastestAddress bool `yaml:"fastest-addr" long:"fastest-addr" description:"Respond to A or AAAA requests only with the fastest IP address" optional:"yes" optional-value:"true"`
+
+	// CacheOptimistic, if set to true, enables the optimistic DNS cache. That
+	// means that cached results will be served even if their cache TTL has
+	// already expired.
+	CacheOptimistic bool `yaml:"cache-optimistic" long:"cache-optimistic" description:"If specified, optimistic DNS cache is enabled" optional:"yes" optional-value:"true"`
+
+	// Cache controls whether DNS responses are cached or not.
+	Cache bool `yaml:"cache" long:"cache" description:"If specified, DNS cache is enabled" optional:"yes" optional-value:"true"`
+
+	// RefuseAny makes the server to refuse requests of type ANY.
+	RefuseAny bool `yaml:"refuse-any" long:"refuse-any" description:"If specified, refuse ANY requests" optional:"yes" optional-value:"true"`
+
+	// EnableEDNSSubnet uses EDNS Client Subnet extension.
+	EnableEDNSSubnet bool `yaml:"edns" long:"edns" description:"Use EDNS Client Subnet extension" optional:"yes" optional-value:"true"`
+
+	// DNS64 defines whether DNS64 functionality is enabled or not.
+	DNS64 bool `yaml:"dns64" long:"dns64" description:"If specified, dnsproxy will act as a DNS64 server" optional:"yes" optional-value:"true"`
+
+	// UsePrivateRDNS makes the server to use private upstreams for reverse DNS
+	// lookups of private addresses, including the requests for authority
+	// records, such as SOA and NS.
+	UsePrivateRDNS bool `yaml:"use-private-rdns" long:"use-private-rdns" description:"If specified, use private upstreams for reverse DNS lookups of private addresses" optional:"yes" optional-value:"true"`
 }
 
 const (
@@ -294,7 +290,11 @@ func run(options *Options) {
 
 	// Prepare the proxy server and its configuration.
 	conf := createProxyConfig(options)
-	dnsProxy := &proxy.Proxy{Config: conf}
+
+	dnsProxy, err := proxy.New(conf)
+	if err != nil {
+		log.Fatalf("creating proxy: %s", err)
+	}
 
 	// Add extra handler if needed.
 	if options.IPv6Disabled {
@@ -303,6 +303,11 @@ func run(options *Options) {
 	}
 
 	// Start the proxy server.
+	//
+	// TODO(e.burkov):  Use signal handler.
+	ctx := context.Background()
+
+	err = dnsProxy.Start(ctx)
 	// rafal code
 	///////////////////////////////////////////////////////////////////////////////
 	proxy.SM.LoadStats("stats.json")
@@ -375,7 +380,7 @@ func run(options *Options) {
 
 	s.Stop()
 	// Stopping the proxy.
-	err = dnsProxy.Stop()
+	err = dnsProxy.Shutdown(ctx)
 	if err != nil {
 		log.Fatalf("cannot stop the DNS proxy due to %s", err)
 	}
@@ -413,8 +418,8 @@ func runPprof(options *Options) {
 }
 
 // createProxyConfig creates proxy.Config from the command line arguments
-func createProxyConfig(options *Options) (conf proxy.Config) {
-	conf = proxy.Config{
+func createProxyConfig(options *Options) (conf *proxy.Config) {
+	conf = &proxy.Config{
 		RatelimitSubnetLenIPv4: options.RatelimitSubnetLenIPv4,
 		RatelimitSubnetLenIPv6: options.RatelimitSubnetLenIPv6,
 
@@ -437,6 +442,8 @@ func createProxyConfig(options *Options) (conf proxy.Config) {
 		UDPBufferSize:          options.UDPBufferSize,
 		HTTPSServerName:        options.HTTPSServerName,
 		MaxGoroutines:          options.MaxGoRoutines,
+		UsePrivateRDNS:         options.UsePrivateRDNS,
+		PrivateSubnets:         netutil.SubnetSetFunc(netutil.IsLocallyServed),
 	}
 
 	if uiStr := options.HTTPSUserinfo; uiStr != "" {
@@ -449,13 +456,13 @@ func createProxyConfig(options *Options) (conf proxy.Config) {
 	}
 
 	// TODO(e.burkov):  Make these methods of [Options].
-	initUpstreams(&conf, options)
-	initEDNS(&conf, options)
-	initBogusNXDomain(&conf, options)
-	initTLSConfig(&conf, options)
-	initDNSCryptConfig(&conf, options)
-	initListenAddrs(&conf, options)
-	initDNS64(&conf, options)
+	initUpstreams(conf, options)
+	initEDNS(conf, options)
+	initBogusNXDomain(conf, options)
+	initTLSConfig(conf, options)
+	initDNSCryptConfig(conf, options)
+	initListenAddrs(conf, options)
+	initSubnets(conf, options)
 
 	return conf
 }
@@ -511,7 +518,7 @@ func initUpstreams(config *proxy.Config, options *Options) {
 	privUpsOpts := &upstream.Options{
 		HTTPVersions: httpVersions,
 		Bootstrap:    boot,
-		Timeout:      mathutil.Min(defaultLocalTimeout, timeout),
+		Timeout:      min(defaultLocalTimeout, timeout),
 	}
 	privUpstreams := loadServersList(options.PrivateRDNSUpstreams)
 
@@ -718,27 +725,33 @@ func initListenAddrs(config *proxy.Config, options *Options) {
 	}
 }
 
-// initDNS64 sets the DNS64 configuration into conf.
-func initDNS64(conf *proxy.Config, options *Options) {
-	if conf.UseDNS64 = options.DNS64; !conf.UseDNS64 {
-		return
-	}
-
-	if conf.PrivateRDNSUpstreamConfig == nil || isEmpty(conf.PrivateRDNSUpstreamConfig) {
-		log.Fatalf("at least one private upstream must be configured to use dns64")
-	}
-
-	var prefs []netip.Prefix
-	for i, p := range options.DNS64Prefix {
+// mustParsePrefixes parses prefixes and considers any error as fatal, logging
+// it with the entity name.
+func mustParsePrefixes(prefixes []string, entity string) (prefs []netip.Prefix) {
+	for i, p := range prefixes {
 		pref, err := netip.ParsePrefix(p)
 		if err != nil {
-			log.Fatalf("parsing prefix at index %d: %v", i, err)
+			log.Fatalf("parsing %s at index %d: %v", entity, i, err)
 		}
 
 		prefs = append(prefs, pref)
 	}
 
-	conf.DNS64Prefs = prefs
+	return prefs
+}
+
+// initSubnets sets the DNS64 configuration into conf.
+func initSubnets(conf *proxy.Config, options *Options) {
+	if conf.UseDNS64 = options.DNS64; conf.UseDNS64 {
+		conf.DNS64Prefs = mustParsePrefixes(options.DNS64Prefix, "dns64 prefix")
+	}
+
+	if options.UsePrivateRDNS {
+		private := mustParsePrefixes(options.PrivateSubnets, "private subnet")
+		if len(private) > 0 {
+			conf.PrivateSubnets = netutil.SliceSubnetSet(private)
+		}
+	}
 }
 
 // IPv6 configuration
