@@ -1,9 +1,8 @@
 package proxy
 
 import (
-	"net/netip"
-
 	"github.com/AdguardTeam/dnsproxy/proxyutil"
+	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/miekg/dns"
 )
 
@@ -16,23 +15,10 @@ func (p *Proxy) isBogusNXDomain(m *dns.Msg) (ok bool) {
 		return false
 	}
 
+	set := netutil.SliceSubnetSet(p.BogusNXDomain)
 	for _, rr := range m.Answer {
 		ip := proxyutil.IPFromRR(rr)
-		if containsIP(p.BogusNXDomain, ip) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func containsIP(nets []netip.Prefix, ip netip.Addr) (ok bool) {
-	if !ip.IsValid() {
-		return false
-	}
-
-	for _, n := range nets {
-		if n.Contains(ip) {
+		if set.Contains(ip) {
 			return true
 		}
 	}
