@@ -8,6 +8,7 @@ import (
 	"net/url"
 	"time"
 
+	"github.com/AdguardTeam/dnsproxy/upstream"
 	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/AdguardTeam/golibs/netutil"
@@ -26,73 +27,54 @@ const (
 	UModeFastestAddr
 )
 
-// BeforeRequestHandler is an optional custom handler called before DNS requests
-// If it returns false, the request won't be processed at all
-type BeforeRequestHandler func(p *Proxy, d *DNSContext) (bool, error)
+// RequestHandler is an optional custom handler for DNS requests.  It's used
+// instead of [Proxy.Resolve] if set.  The resulting error doesn't affect the
+// request processing.  The custom handler is responsible for calling
+// [ResponseHandler], if it doesn't call [Proxy.Resolve].
+//
+// TODO(e.burkov):  Use the same interface-based approach as
+// [BeforeRequestHandler].
+type RequestHandler func(p *Proxy, dctx *DNSContext) (err error)
 
-// RequestHandler is an optional custom handler for DNS requests
-// It is called instead of the default method (Proxy.Resolve())
-// See handler_test.go for examples
-type RequestHandler func(p *Proxy, d *DNSContext) error
-
-// ResponseHandler is a callback method that is called when DNS query has been processed
-// d -- current DNS query context (contains response if it was successful)
-// err -- error (if any)
-type ResponseHandler func(d *DNSContext, err error)
+// ResponseHandler is an optional custom handler called when DNS query has been
+// processed.  When called from [Proxy.Resolve], dctx will contain the response
+// message if the upstream or cache succeeded.  err is only not nil if the
+// upstream failed to respond.
+//
+// TODO(e.burkov):  Use the same interface-based approach as
+// [BeforeRequestHandler].
+type ResponseHandler func(dctx *DNSContext, err error)
 
 // Config contains all the fields necessary for proxy configuration
 //
 // TODO(a.garipov): Consider extracting conf blocks for better fieldalignment.
 type Config struct {
-	// Listeners
-	// --
-
-	UDPListenAddr         []*net.UDPAddr // if nil, then it does not listen for UDP
-	TCPListenAddr         []*net.TCPAddr // if nil, then it does not listen for TCP
-	HTTPSListenAddr       []*net.TCPAddr // if nil, then it does not listen for HTTPS (DoH)
-	TLSListenAddr         []*net.TCPAddr // if nil, then it does not listen for TLS (DoT)
-	QUICListenAddr        []*net.UDPAddr // if nil, then it does not listen for QUIC (DoQ)
-	DNSCryptUDPListenAddr []*net.UDPAddr // if nil, then it does not listen for DNSCrypt
-	DNSCryptTCPListenAddr []*net.TCPAddr // if nil, then it does not listen for DNSCrypt
-
-	// Encryption configuration
-	// --
-
-	TLSConfig            *tls.Config    // necessary for TLS, HTTPS, QUIC
-	HTTP3                bool           // if true, HTTPS server will also support HTTP/3
-	DNSCryptProviderName string         // DNSCrypt provider name
-	DNSCryptResolverCert *dnscrypt.Cert // DNSCrypt resolver certificate
-
-	// Rate-limiting and anti-DNS amplification measures
-	// --
-	//
-	// TODO(s.chzhen):  Extract ratelimit settings to a separate structure.
-
-	// RatelimitSubnetLenIPv4 is a subnet length for IPv4 addresses used for
-	// rate limiting requests.
-	RatelimitSubnetLenIPv4 int
-
-	// RatelimitSubnetLenIPv6 is a subnet length for IPv6 addresses used for
-	// rate limiting requests.
-	RatelimitSubnetLenIPv6 int
-
-	// Ratelimit is a maximum number of requests per second from a given IP (0
-	// to disable).
-	Ratelimit int
-
-	// RatelimitWhitelist is a list of IP addresses excluded from rate limiting.
-	RatelimitWhitelist []netip.Addr
-
-	// RefuseAny makes proxy refuse the requests of type ANY.
-	RefuseAny bool
-
 	// TrustedProxies is the trusted list of CIDR networks to detect proxy
 	// servers addresses from where the DoH requests should be handled.  The
 	// value of nil makes Proxy not trust any address.
 	TrustedProxies netutil.SubnetSet
 
-	// Upstream DNS servers and their settings
-	// --
+	// PrivateSubnets is the set of private networks.  Client having an address
+	// within this set is able to resolve PTR requests for addresses within this
+	// set.
+	PrivateSubnets netutil.SubnetSet
+
+	// MessageConstructor used to build DNS messages.  If nil, the default
+	// constructor will be used.
+	MessageConstructor MessageConstructor
+
+	// BeforeRequestHandler is an optional custom handler called before each DNS
+	// request is started processing, see [BeforeRequestHandler].  The default
+	// no-op implementation is used, if it's nil.
+	BeforeRequestHandler BeforeRequestHandler
+
+	// RequestHandler is an optional custom handler for DNS requests.  It's used
+	// instead of [Proxy.Resolve] if set.  See [RequestHandler].
+	RequestHandler RequestHandler
+
+	// ResponseHandler is an optional custom handler called when DNS query has
+	// been processed.  See [ResponseHandler].
+	ResponseHandler ResponseHandler
 
 	// UpstreamConfig is a general set of DNS servers to forward requests to.
 	UpstreamConfig *UpstreamConfig
@@ -107,6 +89,105 @@ type Config struct {
 	// general set fails responding.
 	Fallbacks *UpstreamConfig
 
+	// Userinfo is the sole permitted userinfo for the DoH basic authentication.
+	// If Userinfo is set, all DoH queries are required to have this basic
+	// authentication information.
+	Userinfo *url.Userinfo
+
+	// TLSConfig is the TLS configuration.  Required for DNS-over-TLS,
+	// DNS-over-HTTP, and DNS-over-QUIC servers.
+	TLSConfig *tls.Config
+
+	// DNSCryptResolverCert is the DNSCrypt resolver certificate.  Required for
+	// DNSCrypt server.
+	DNSCryptResolverCert *dnscrypt.Cert
+
+	// DNSCryptProviderName is the DNSCrypt provider name.  Required for
+	// DNSCrypt server.
+	DNSCryptProviderName string
+
+	// HTTPSServerName sets the Server header of the HTTPS server responses, if
+	// not empty.
+	HTTPSServerName string
+
+	// UDPListenAddr is the set of UDP addresses to listen for plain
+	// DNS-over-UDP requests.
+	UDPListenAddr []*net.UDPAddr
+
+	// TCPListenAddr is the set of TCP addresses to listen for plain
+	// DNS-over-TCP requests.
+	TCPListenAddr []*net.TCPAddr
+
+	// HTTPSListenAddr is the set of TCP addresses to listen for DNS-over-HTTPS
+	// requests.
+	HTTPSListenAddr []*net.TCPAddr
+
+	// TLSListenAddr is the set of TCP addresses to listen for DNS-over-TLS
+	// requests.
+	TLSListenAddr []*net.TCPAddr
+
+	// QUICListenAddr is the set of UDP addresses to listen for DNS-over-QUIC
+	// requests.
+	QUICListenAddr []*net.UDPAddr
+
+	// DNSCryptUDPListenAddr is the set of UDP addresses to listen for DNSCrypt
+	// requests.
+	DNSCryptUDPListenAddr []*net.UDPAddr
+
+	// DNSCryptTCPListenAddr is the set of TCP addresses to listen for DNSCrypt
+	// requests.
+	DNSCryptTCPListenAddr []*net.TCPAddr
+
+	// BogusNXDomain is the set of networks used to transform responses into
+	// NXDOMAIN ones if they contain at least a single IP address within these
+	// networks.  It's similar to dnsmasq's "bogus-nxdomain".
+	BogusNXDomain []netip.Prefix
+
+	// DNS64Prefs is the set of NAT64 prefixes used for DNS64 handling.  nil
+	// value disables the feature.  An empty value will be interpreted as the
+	// default Well-Known Prefix.
+	DNS64Prefs []netip.Prefix
+
+	// RatelimitWhitelist is a list of IP addresses excluded from rate limiting.
+	RatelimitWhitelist []netip.Addr
+
+	// EDNSAddr is the ECS IP used in request.
+	EDNSAddr net.IP
+
+	// TODO(s.chzhen):  Extract ratelimit settings to a separate structure.
+
+	// RatelimitSubnetLenIPv4 is a subnet length for IPv4 addresses used for
+	// rate limiting requests.
+	RatelimitSubnetLenIPv4 int
+
+	// RatelimitSubnetLenIPv6 is a subnet length for IPv6 addresses used for
+	// rate limiting requests.
+	RatelimitSubnetLenIPv6 int
+
+	// Ratelimit is a maximum number of requests per second from a given IP (0
+	// to disable).
+	Ratelimit int
+
+	// CacheSizeBytes is the maximum cache size in bytes.
+	CacheSizeBytes int
+
+	// CacheMinTTL is the minimum TTL for cached DNS responses in seconds.
+	CacheMinTTL uint32
+
+	// CacheMaxTTL is the maximum TTL for cached DNS responses in seconds.
+	CacheMaxTTL uint32
+
+	// MaxGoroutines is the maximum number of goroutines processing DNS
+	// requests.  Important for mobile users.
+	//
+	// TODO(a.garipov): Rename this to something like “MaxDNSRequestGoroutines”
+	// in a later major version, as it doesn't actually limit all goroutines.
+	MaxGoroutines uint
+
+	// The size of the read buffer on the underlying socket.  Larger read
+	// buffers can handle larger bursts of requests before packets get dropped.
+	UDPBufferSize int
+
 	// UpstreamMode determines the logic through which upstreams will be used.
 	UpstreamMode UpstreamModeType
 
@@ -115,10 +196,11 @@ type Config struct {
 	// value will be replaced with the default one.
 	FastestPingTimeout time.Duration
 
-	// BogusNXDomain is the set of networks used to transform responses into
-	// NXDOMAIN ones if they contain at least a single IP address within these
-	// networks.  It's similar to dnsmasq's "bogus-nxdomain".
-	BogusNXDomain []netip.Prefix
+	// RefuseAny makes proxy refuse the requests of type ANY.
+	RefuseAny bool
+
+	// HTTP3 enables HTTP/3 support for HTTPS server.
+	HTTP3 bool
 
 	// Enable EDNS Client Subnet option DNS requests to the upstream server will
 	// contain an OPT record with Client Subnet option.  If the original request
@@ -142,61 +224,24 @@ type Config struct {
 	// never be used for clients with public IP addresses.
 	EnableEDNSClientSubnet bool
 
-	// EDNSAddr is the ECS IP used in request.
-	EDNSAddr net.IP
+	// CacheEnabled defines if the response cache should be used.
+	CacheEnabled bool
 
-	// Cache settings
-	// --
-
-	CacheEnabled   bool   // cache status
-	CacheSizeBytes int    // Cache size (in bytes). Default: 64k
-	CacheMinTTL    uint32 // Minimum TTL for DNS entries (in seconds).
-	CacheMaxTTL    uint32 // Maximum TTL for DNS entries (in seconds).
-	// CacheOptimistic defines if the optimistic cache mechanism should be
-	// used.
+	// CacheOptimistic defines if the optimistic cache mechanism should be used.
 	CacheOptimistic bool
-
-	// Handlers (for the case when dnsproxy is used as a library)
-	// --
-
-	BeforeRequestHandler BeforeRequestHandler // callback that is called before each request
-	RequestHandler       RequestHandler       // callback that can handle incoming DNS requests
-	ResponseHandler      ResponseHandler      // response callback
-
-	// Other settings
-	// --
-
-	// HTTPSServerName sets the Server header of the HTTPS server responses, if
-	// not empty.
-	HTTPSServerName string
-
-	// Userinfo is the sole permitted userinfo for the DoH basic authentication.
-	// If Userinfo is set, all DoH queries are required to have this basic
-	// authentication information.
-	Userinfo *url.Userinfo
-
-	// MaxGoroutines is the maximum number of goroutines processing DNS
-	// requests.  Important for mobile users.
-	//
-	// TODO(a.garipov): Rename this to something like
-	// “MaxDNSRequestGoroutines” in a later major version, as it doesn't
-	// actually limit all goroutines.
-	MaxGoroutines uint
-
-	// The size of the read buffer on the underlying socket. Larger read buffers can handle
-	// larger bursts of requests before packets get dropped.
-	UDPBufferSize int
 
 	// UseDNS64 enables DNS64 handling.  If true, proxy will translate IPv4
 	// answers into IPv6 answers using first of DNS64Prefs.  Note also that PTR
 	// requests for addresses within the specified networks are considered
 	// private and will be forwarded as PrivateRDNSUpstreamConfig specifies.
+	// Those will be responded with NXDOMAIN if UsePrivateRDNS is false.
 	UseDNS64 bool
 
-	// DNS64Prefs is the set of NAT64 prefixes used for DNS64 handling.  nil
-	// value disables the feature.  An empty value will be interpreted as the
-	// default Well-Known Prefix.
-	DNS64Prefs []netip.Prefix
+	// UsePrivateRDNS defines if the PTR requests for private IP addresses
+	// should be resolved via PrivateRDNSUpstreamConfig.  Note that it requires
+	// a valid PrivateRDNSUpstreamConfig with at least a single general upstream
+	// server.
+	UsePrivateRDNS bool
 
 	// PreferIPv6 tells the proxy to prefer IPv6 addresses when bootstrapping
 	// upstreams that use hostnames.
@@ -205,28 +250,23 @@ type Config struct {
 
 // validateConfig verifies that the supplied configuration is valid and returns
 // an error if it's not.
-func (p *Proxy) validateConfig() error {
-	err := p.validateListenAddrs()
-	if err != nil {
-		// Don't wrap the error since it's informative enough as is.
-		return err
-	}
-
+func (p *Proxy) validateConfig() (err error) {
 	err = p.UpstreamConfig.validate()
 	if err != nil {
 		return fmt.Errorf("validating general upstreams: %w", err)
 	}
 
-	// Allow both [Proxy.PrivateRDNSUpstreamConfig] and [Proxy.Fallbacks] to be
-	// nil, but not empty.  nil means using the default values for those.
-
-	err = p.PrivateRDNSUpstreamConfig.validate()
-	if err != nil && !errors.Is(err, errNoDefaultUpstreams) {
-		return fmt.Errorf("validating private RDNS upstreams: %w", err)
+	err = ValidatePrivateConfig(p.PrivateRDNSUpstreamConfig, p.privateNets)
+	if err != nil {
+		if p.UsePrivateRDNS || errors.Is(err, upstream.ErrNoUpstreams) {
+			return fmt.Errorf("validating private RDNS upstreams: %w", err)
+		}
 	}
 
+	// Allow [Proxy.Fallbacks] to be nil, but not empty.  nil means not to use
+	// fallbacks at all.
 	err = p.Fallbacks.validate()
-	if err != nil && !errors.Is(err, errNoDefaultUpstreams) {
+	if errors.Is(err, upstream.ErrNoUpstreams) {
 		return fmt.Errorf("validating fallbacks: %w", err)
 	}
 
@@ -289,7 +329,7 @@ func (p *Proxy) logConfigInfo() {
 	}
 
 	if p.RefuseAny {
-		log.Info("The server is configured to refuse ANY requests")
+		log.Info("dnsproxy: server will refuse requests of type ANY")
 	}
 
 	if len(p.BogusNXDomain) > 0 {
