@@ -1,55 +1,65 @@
 package proxy
 
 import (
+	"context"
+	"net"
 	"sync"
 	"testing"
 	"time"
 
+	"github.com/AdguardTeam/golibs/testutil"
 	"github.com/miekg/dns"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestFilteringHandler(t *testing.T) {
 	// Initializing the test middleware
-	m := sync.RWMutex{}
+	m := &sync.RWMutex{}
 	blockResponse := false
 
 	// Prepare the proxy server
-	dnsProxy := createTestProxy(t, nil)
-	dnsProxy.RequestHandler = func(p *Proxy, d *DNSContext) error {
-		m.Lock()
-		defer m.Unlock()
+	dnsProxy := mustNew(t, &Config{
+		UDPListenAddr:          []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TCPListenAddr:          []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		UpstreamConfig:         newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr),
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+		RequestHandler: func(p *Proxy, d *DNSContext) error {
+			m.Lock()
+			defer m.Unlock()
 
-		if !blockResponse {
-			// Use the default Resolve method if response is not blocked
-			return p.Resolve(d)
-		}
+			if !blockResponse {
+				// Use the default Resolve method if response is not blocked
+				return p.Resolve(d)
+			}
 
-		resp := dns.Msg{}
-		resp.SetRcode(d.Req, dns.RcodeNotImplemented)
-		resp.RecursionAvailable = true
+			resp := dns.Msg{}
+			resp.SetRcode(d.Req, dns.RcodeNotImplemented)
+			resp.RecursionAvailable = true
 
-		// Set the response right away
-		d.Res = &resp
-		return nil
-	}
+			// Set the response right away
+			d.Res = &resp
+			return nil
+		},
+	})
 
 	// Start listening
-	err := dnsProxy.Start()
-	if err != nil {
-		t.Fatalf("cannot start the DNS proxy: %s", err)
-	}
+	ctx := context.Background()
+	err := dnsProxy.Start(ctx)
+	require.NoError(t, err)
+	testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
 
 	// Create a DNS-over-UDP client connection
 	addr := dnsProxy.Addr(ProtoUDP)
 	client := &dns.Client{Net: "udp", Timeout: 500 * time.Millisecond}
 
 	// Send the first message (not blocked)
-	req := createTestMessage()
+	req := newTestMessage()
 
 	r, _, err := client.Exchange(req, addr.String())
-	if err != nil {
-		t.Fatalf("error in the first request: %s", err)
-	}
+	require.NoError(t, err)
 	requireResponse(t, req, r)
 
 	// Now send the second and make sure it is blocked
@@ -58,16 +68,6 @@ func TestFilteringHandler(t *testing.T) {
 	m.Unlock()
 
 	r, _, err = client.Exchange(req, addr.String())
-	if err != nil {
-		t.Fatalf("error in the second request: %s", err)
-	}
-	if r.Rcode != dns.RcodeNotImplemented {
-		t.Fatalf("second request was not blocked")
-	}
-
-	// Stop the proxy
-	err = dnsProxy.Stop()
-	if err != nil {
-		t.Fatalf("cannot stop the DNS proxy: %s", err)
-	}
+	require.NoError(t, err)
+	assert.Equal(t, dns.RcodeNotImplemented, r.Rcode)
 }
