@@ -11,7 +11,9 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/AdguardTeam/golibs/errors"
 	"github.com/AdguardTeam/golibs/log"
+	"github.com/AdguardTeam/golibs/netutil"
 	"github.com/miekg/dns"
 )
 
@@ -95,6 +97,11 @@ func (p *Proxy) startListeners(ctx context.Context) error {
 	return nil
 }
 
+// handleDNSRequest processes the context.  The only error it returns is the one
+// from the [RequestHandler], or [Resolve] if the [RequestHandler] is not set.
+// d is left without a response as the documentation to [BeforeRequestHandler]
+// says, and if it's ratelimited.
+func (p *Proxy) handleDNSRequest(d *DNSContext) (err error) {
 // handleDNSRequest processes the incoming packet bytes and returns with an optional response packet.
 func (p *Proxy) handleDNSRequest(d *DNSContext) error {
 	// rafal
@@ -104,57 +111,36 @@ func (p *Proxy) handleDNSRequest(d *DNSContext) error {
 	p.logDNSMessage(d.Req)
 
 	if d.Req.Response {
+		log.Debug("dnsproxy: dropping incoming response packet from %s", d.Addr)
+
 		//log.Debug("Dropping incoming Reply packet from %s", d.Addr.String())
 		return nil
 	}
 
-	if p.BeforeRequestHandler != nil {
-		ok, err := p.BeforeRequestHandler(p, d)
-		if err != nil {
-			//log.Error("Error in the BeforeRequestHandler: %s", err)	// rafal
-			d.Res = p.genServerFailure(d.Req)
-			p.respond(d)
-			return nil
-		}
-		if !ok {
-			return nil // do nothing, don't reply
-		}
+	ip := d.Addr.Addr()
+	d.IsPrivateClient = p.privateNets.Contains(ip)
+
+	if !p.handleBefore(d) {
+		return nil
 	}
 
 	// ratelimit based on IP only, protects CPU cycles and outbound connections
-	if d.Proto == ProtoUDP && p.isRatelimited(d.Addr.Addr()) {
-		//log.Tracef("Ratelimiting %v based on IP only", d.Addr)
-		return nil // do nothing, don't reply, we got ratelimited
+	//
+	// TODO(e.burkov):  Investigate if written above true and move to UDP server
+	// implementation?
+	if d.Proto == ProtoUDP && p.isRatelimited(ip) {
+		log.Debug("dnsproxy: ratelimiting %s based on IP only", d.Addr)
+
+		// Don't reply to ratelimitted clients.
+		return nil
 	}
 
-	if len(d.Req.Question) != 1 {
-		//log.Debug("got invalid number of questions: %v", len(d.Req.Question))	// rafal
-		d.Res = p.genServerFailure(d.Req)
-	}
-
-	// refuse ANY requests (anti-DDOS measure)
-	if p.RefuseAny && len(d.Req.Question) > 0 && d.Req.Question[0].Qtype == dns.TypeANY {
-		log.Tracef("Refusing type=ANY request")
-		d.Res = p.genNotImpl(d.Req)
-	}
-
-	var err error
-
+	d.Res = p.validateRequest(d)
 	if d.Res == nil {
-		if len(p.UpstreamConfig.Upstreams) == 0 {
-			panic("SHOULD NOT HAPPEN: no default upstreams specified")
-		}
-
-		// execute the DNS request
-		// if there is a custom middleware configured, use it
 		if p.RequestHandler != nil {
-			err = p.RequestHandler(p, d)
+			err = errors.Annotate(p.RequestHandler(p, d), "using request handler: %w")
 		} else {
-			err = p.Resolve(d)
-		}
-
-		if err != nil {
-			err = fmt.Errorf("talking to dns upstream: %w", err)
+			err = errors.Annotate(p.Resolve(d), "using default request handler: %w")
 		}
 	}
 
@@ -165,6 +151,66 @@ func (p *Proxy) handleDNSRequest(d *DNSContext) error {
 	p.respond(d)
 
 	return err
+}
+
+// validateRequest returns a response for invalid request or nil if the request
+// is ok.
+func (p *Proxy) validateRequest(d *DNSContext) (resp *dns.Msg) {
+	switch {
+	case len(d.Req.Question) != 1:
+		log.Debug("dnsproxy: got invalid number of questions: %d", len(d.Req.Question))
+
+		// TODO(e.burkov):  Probably, FORMERR would be a better choice here.
+		// Check out RFC.
+		return p.messages.NewMsgSERVFAIL(d.Req)
+	case p.RefuseAny && d.Req.Question[0].Qtype == dns.TypeANY:
+		// Refuse requests of type ANY (anti-DDOS measure).
+		log.Debug("dnsproxy: refusing type=ANY request")
+
+		return p.messages.NewMsgNOTIMPLEMENTED(d.Req)
+	case p.recDetector.check(d.Req):
+		log.Debug("dnsproxy: recursion detected resolving %q", d.Req.Question[0].Name)
+
+		return p.messages.NewMsgNXDOMAIN(d.Req)
+	case d.isForbiddenARPA(p.privateNets):
+		log.Debug("dnsproxy: %s requests a private arpa domain %q", d.Addr, d.Req.Question[0].Name)
+
+		return p.messages.NewMsgNXDOMAIN(d.Req)
+	default:
+		return nil
+	}
+}
+
+// isForbiddenARPA returns true if dctx contains a PTR, SOA, or NS request for
+// some private address and client's address is not within the private network.
+// Otherwise, it sets [DNSContext.RequestedPrivateRDNS] for future use.
+func (dctx *DNSContext) isForbiddenARPA(privateNets netutil.SubnetSet) (ok bool) {
+	q := dctx.Req.Question[0]
+	switch q.Qtype {
+	case dns.TypePTR, dns.TypeSOA, dns.TypeNS:
+		// Go on.
+		//
+		// TODO(e.burkov):  Reconsider the list of types involved to private
+		// address space.  Perhaps, use the logic for any type.  See
+		// https://www.rfc-editor.org/rfc/rfc6761.html#section-6.1.
+	default:
+		return false
+	}
+
+	requestedPref, err := netutil.ExtractReversedAddr(q.Name)
+	if err != nil {
+		log.Debug("proxy: parsing reversed subnet: %v", err)
+
+		return false
+	}
+
+	if privateNets.Contains(requestedPref.Addr()) {
+		dctx.RequestedPrivateRDNS = requestedPref
+
+		return !dctx.IsPrivateClient
+	}
+
+	return false
 }
 
 // respond writes the specified response to the client (or does nothing if d.Res is empty)
@@ -209,27 +255,6 @@ func (p *Proxy) setMinMaxTTL(r *dns.Msg) {
 			rr.Header().Ttl = newTTL
 		}
 	}
-}
-
-func (p *Proxy) genServerFailure(request *dns.Msg) *dns.Msg {
-	return p.genWithRCode(request, dns.RcodeServerFailure)
-}
-
-func (p *Proxy) genNotImpl(request *dns.Msg) (resp *dns.Msg) {
-	resp = p.genWithRCode(request, dns.RcodeNotImplemented)
-	// NOTIMPL without EDNS is treated as 'we don't support EDNS', so
-	// explicitly set it.
-	resp.SetEdns0(1452, false)
-
-	return resp
-}
-
-func (p *Proxy) genWithRCode(req *dns.Msg, code int) (resp *dns.Msg) {
-	resp = &dns.Msg{}
-	resp.SetRcode(req, code)
-	resp.RecursionAvailable = true
-
-	return resp
 }
 
 func (p *Proxy) logDNSMessage(m *dns.Msg) {
