@@ -195,8 +195,7 @@ func (p *Proxy) checkBasicAuth(
 	log.Error("dnsproxy: basic auth failed for user %q from raddr %s", user, raddr)
 
 	h := w.Header()
-	// TODO(a.garipov): Add to httphdr.
-	h.Set("Www-Authenticate", `Basic realm="DNS", charset="UTF-8"`)
+	h.Set(httphdr.WWWAuthenticate, `Basic realm="DNS", charset="UTF-8"`)
 	http.Error(w, "Authorization required", http.StatusUnauthorized)
 
 	return false
