@@ -17,14 +17,7 @@ import (
 )
 
 func TestQuicProxy(t *testing.T) {
-	// Prepare the proxy server.
-	serverConfig, caPem := createServerTLSConfig(t)
-	dnsProxy := createTestProxy(t, serverConfig)
-	testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
-
-	// Start listening.
-	err := dnsProxy.Start()
-	require.NoError(t, err)
+	serverConfig, caPem := newTLSConfig(t)
 
 	roots := x509.NewCertPool()
 	roots.AppendCertsFromPEM(caPem)
@@ -34,54 +27,99 @@ func TestQuicProxy(t *testing.T) {
 		NextProtos: append([]string{NextProtoDQ}, compatProtoDQ...),
 	}
 
-	// Create a DNS-over-QUIC client connection.
-	addr := dnsProxy.Addr(ProtoQUIC)
+	conf := &Config{
+		QUICListenAddr:         []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TLSConfig:              serverConfig,
+		UpstreamConfig:         newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr),
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+	}
 
-	// Open a QUIC connection.
-	conn, err := quic.DialAddrEarly(context.Background(), addr.String(), tlsConfig, nil)
-	require.NoError(t, err)
-	testutil.CleanupAndRequireSuccess(t, func() (err error) {
-		return conn.CloseWithError(DoQCodeNoError, "")
+	var addr *net.UDPAddr
+	t.Run("run", func(t *testing.T) {
+		dnsProxy := mustNew(t, conf)
+
+		ctx := context.Background()
+		err := dnsProxy.Start(ctx)
+		require.NoError(t, err)
+		testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
+
+		addr = testutil.RequireTypeAssert[*net.UDPAddr](t, dnsProxy.Addr(ProtoQUIC))
+
+		conn, err := quic.DialAddrEarly(context.Background(), addr.String(), tlsConfig, nil)
+		require.NoError(t, err)
+		testutil.CleanupAndRequireSuccess(t, func() (err error) {
+			return conn.CloseWithError(DoQCodeNoError, "")
+		})
+
+		for range 10 {
+			sendTestQUICMessage(t, conn, DoQv1)
+
+			// Send a message encoded for a draft version as well.
+			sendTestQUICMessage(t, conn, DoQv1Draft)
+		}
 	})
+	require.False(t, t.Failed())
 
-	// Send several test messages.
-	for i := 0; i < 10; i++ {
+	conf.QUICListenAddr = []*net.UDPAddr{addr}
+	conf.UpstreamConfig = newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr)
+
+	t.Run("rerun", func(t *testing.T) {
+		dnsProxy := mustNew(t, conf)
+
+		ctx := context.Background()
+		err := dnsProxy.Start(ctx)
+		require.NoError(t, err)
+		testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
+
+		conn, err := quic.DialAddrEarly(context.Background(), addr.String(), tlsConfig, nil)
+		require.NoError(t, err)
+		testutil.CleanupAndRequireSuccess(t, func() (err error) {
+			return conn.CloseWithError(DoQCodeNoError, "")
+		})
+
 		sendTestQUICMessage(t, conn, DoQv1)
 
 		// Send a message encoded for a draft version as well.
 		sendTestQUICMessage(t, conn, DoQv1Draft)
-	}
+	})
 }
 
 func TestQuicProxy_largePackets(t *testing.T) {
-	// Prepare the proxy server.
-	serverConfig, caPem := createServerTLSConfig(t)
-	dnsProxy := createTestProxy(t, serverConfig)
-
-	// Make sure the request does not go to any real upstream.
-	dnsProxy.RequestHandler = func(p *Proxy, d *DNSContext) (err error) {
-		resp := &dns.Msg{}
-		resp.SetReply(d.Req)
-		resp.Answer = []dns.RR{
-			&dns.A{
+	serverConfig, caPem := newTLSConfig(t)
+	dnsProxy := mustNew(t, &Config{
+		TLSListenAddr:          []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		HTTPSListenAddr:        []*net.TCPAddr{net.TCPAddrFromAddrPort(localhostAnyPort)},
+		QUICListenAddr:         []*net.UDPAddr{net.UDPAddrFromAddrPort(localhostAnyPort)},
+		TLSConfig:              serverConfig,
+		UpstreamConfig:         newTestUpstreamConfig(t, defaultTimeout, testDefaultUpstreamAddr),
+		TrustedProxies:         defaultTrustedProxies,
+		RatelimitSubnetLenIPv4: 24,
+		RatelimitSubnetLenIPv6: 64,
+		// Make sure the request does not go to any real upstream.
+		RequestHandler: func(_ *Proxy, d *DNSContext) (err error) {
+			resp := &dns.Msg{}
+			resp.SetReply(d.Req)
+			resp.Answer = []dns.RR{&dns.A{
 				Hdr: dns.RR_Header{
 					Name:   d.Req.Question[0].Name,
 					Rrtype: dns.TypeA,
 					Class:  dns.ClassINET,
 				},
 				A: net.IP{8, 8, 8, 8},
-			},
-		}
-		d.Res = resp
+			}}
+			d.Res = resp
 
-		return nil
-	}
-
-	testutil.CleanupAndRequireSuccess(t, dnsProxy.Stop)
+			return nil
+		},
+	})
 
 	// Start listening.
-	err := dnsProxy.Start()
+	ctx := context.Background()
+	err := dnsProxy.Start(ctx)
 	require.NoError(t, err)
+	testutil.CleanupAndRequireSuccess(t, func() (err error) { return dnsProxy.Shutdown(ctx) })
 
 	roots := x509.NewCertPool()
 	roots.AppendCertsFromPEM(caPem)
@@ -102,7 +140,7 @@ func TestQuicProxy_largePackets(t *testing.T) {
 	})
 
 	// Create a test message large enough to take multiple QUIC frames.
-	msg := createTestMessage()
+	msg := newTestMessage()
 	msg.Extra = []dns.RR{
 		&dns.OPT{
 			Hdr: dns.RR_Header{Name: ".", Rrtype: dns.TypeOPT, Class: 4096},
@@ -193,7 +231,7 @@ func writeQUICStream(buf []byte, stream quic.Stream) (err error) {
 
 // sendTestQUICMessage send a test message to the specified QUIC connection.
 func sendTestQUICMessage(t *testing.T, conn quic.Connection, doqVersion DoQVersion) {
-	msg := createTestMessage()
+	msg := newTestMessage()
 	resp := sendQUICMessage(t, msg, conn, doqVersion)
 	requireResponse(t, msg, resp)
 }
