@@ -99,8 +99,6 @@ func newDoH(addr *url.URL, opts *Options) (u Upstream, err error) {
 			Tracer:          opts.QUICTracer,
 		},
 		quicConfMu: &sync.Mutex{},
-		// #nosec G402 -- TLS certificate verification could be disabled by
-		// configuration.
 		tlsConf: &tls.Config{
 			ServerName:   addr.Hostname(),
 			RootCAs:      opts.RootCAs,
@@ -108,8 +106,10 @@ func newDoH(addr *url.URL, opts *Options) (u Upstream, err error) {
 			// Use the default capacity for the LRU cache.  It may be useful to
 			// store several caches since the user may be routed to different
 			// servers in case there's load balancing on the server-side.
-			ClientSessionCache:    tls.NewLRUClientSessionCache(0),
-			MinVersion:            tls.VersionTLS12,
+			ClientSessionCache: tls.NewLRUClientSessionCache(0),
+			MinVersion:         tls.VersionTLS12,
+			// #nosec G402 -- TLS certificate verification could be disabled by
+			// configuration.
 			InsecureSkipVerify:    opts.InsecureSkipVerify,
 			VerifyPeerCertificate: opts.VerifyServerCertificate,
 			VerifyConnection:      opts.VerifyConnection,
@@ -555,7 +555,7 @@ func (p *dnsOverHTTPS) createTransportH3(
 		},
 		DisableCompression: true,
 		TLSClientConfig:    tlsConfig,
-		QuicConfig:         p.getQUICConfig(),
+		QUICConfig:         p.getQUICConfig(),
 	}
 
 	return &http3Transport{baseTransport: rt}, nil
