@@ -54,8 +54,6 @@ func newDoT(addr *url.URL, opts *Options) (ups Upstream, err error) {
 	tlsUps := &dnsOverTLS{
 		addr:      addr,
 		getDialer: newDialerInitializer(addr, opts),
-		// #nosec G402 -- TLS certificate verification could be disabled by
-		// configuration.
 		tlsConf: &tls.Config{
 			ServerName:   addr.Hostname(),
 			RootCAs:      opts.RootCAs,
@@ -63,8 +61,10 @@ func newDoT(addr *url.URL, opts *Options) (ups Upstream, err error) {
 			// Use the default capacity for the LRU cache.  It may be useful to
 			// store several caches since the user may be routed to different
 			// servers in case there's load balancing on the server-side.
-			ClientSessionCache:    tls.NewLRUClientSessionCache(0),
-			MinVersion:            tls.VersionTLS12,
+			ClientSessionCache: tls.NewLRUClientSessionCache(0),
+			MinVersion:         tls.VersionTLS12,
+			// #nosec G402 -- TLS certificate verification could be disabled by
+			// configuration.
 			InsecureSkipVerify:    opts.InsecureSkipVerify,
 			VerifyPeerCertificate: opts.VerifyServerCertificate,
 			VerifyConnection:      opts.VerifyConnection,
